refactor: 안쓰는 폴더 지우기
