--- src/components/UserSidebar/index.jsx
@@ -1,14 +1,0 @@
-import { UserSidebarContainer } from './style';
-import UserList from './UserList';
-import UserSearchBar from './UserSearchBar';
-
-const UserSidebar = () => {
-  return (
-    <UserSidebarContainer>
-      <UserSearchBar />
-      <UserList />
-    </UserSidebarContainer>
-  );
-};
-
-export default UserSidebar;
