--- src/components/UserSidebar/style.js
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const UserSidebarContainer = styled.div`
-  width: 280px;
-  border-left: 1px solid rgba(0, 0, 0, 0.1);
-  overflow-y: scroll;
-`;
